backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
Pillow>=10.2.0
numpy>=1.26.0
opencv-python-headless>=4.9.0
aiofiles>=23.2.1
